changing data type for tokens and xp balance

+++ Assets/src/Runtime/APIClient/ResponseData/ResponseData.cs
@@ -75,8 +75,8 @@ namespace TA.APIClient.ResponseData{
         public int gameCoin;
         public int freeCredits;
         public int boughtCredits;
-        public int tokens;
-        public int xpBalance;
+        public float tokens;
+        public float xpBalance;
         public float native;
     }
